IBX-10679: Extract Field abstract class
